+++ max_div/solver/_strategies/_initialization/_base.py
@@ -24,11 +24,11 @@ class InitializationStrategy(StrategyBase, ABC):
     #  Factory Methods
     # -------------------------------------------------------------------------
     @classmethod
-    def one_shot_random(cls, constrained: bool = True, uniform: bool = False) -> Self:
-        """Create a InitOneShotRandom initialization strategy."""
-        from ._init_one_shot_random import InitOneShotRandom
+    def random_one_shot(cls, constrained: bool = True, uniform: bool = False) -> Self:
+        """Create a InitRandomOneShot initialization strategy."""
+        from ._init_random_one_shot import InitRandomOneShot
 
-        return InitOneShotRandom(
+        return InitRandomOneShot(
             constrained=constrained,
             uniform=uniform,
         )